Move AnalyticsParams to analytics module
COSDK-500
