+++ core/src/main/java/com/adyen/checkout/core/internal/ui/model/CommonComponentParams.kt
@@ -10,6 +10,7 @@ package com.adyen.checkout.core.internal.ui.model
 
 import androidx.annotation.RestrictTo
 import com.adyen.checkout.core.Environment
+import com.adyen.checkout.core.analytics.internal.AnalyticsParams
 import com.adyen.checkout.core.data.model.Amount
 import java.util.Locale
